refactor(core): move generators out of `compiler`
