+++ packages/atropa/src/core/generators/block.ts
@@ -1,5 +1,9 @@
 import { deepMerge, objectMap } from '@antfu/utils'
 import { ensureNamespaces } from '../utils'
+import type {
+	BlockTemplate,
+	BlockFormatVersion
+} from '../../schema/atropa/server/block'
 
 interface UserTemplate {
 	namespace?: string
@@ -15,9 +19,31 @@ interface VanillaTemplate {
 	components?: Record<string, any>
 	events?: Record<string, any>
 }
-interface BlockObject {
+interface Block {
 	format_version: string
 	'minecraft:block': VanillaTemplate
+}
+
+/**
+ * # Define Block
+ *
+ * Generates a new block based on the given templates.
+ * @param version The format version of the block.
+ * @param fn A callback function with function parameters used to define the block.
+ * @returns A block.
+ */
+export function defineBlock<Version extends BlockFormatVersion>(
+	version: Version,
+	fn: (template: BlockTemplate<Version>) => void
+): Record<string, any> {
+	try {
+		const template = {}
+
+		fn(processTemplate(template) as BlockTemplate<Version>)
+		return transformTemplate(template, version)
+	} catch (error) {
+		throw new Error(`Failed to transform block template`, error as Error)
+	}
 }
 
 export function processTemplate(template: VanillaTemplate): UserTemplate {
@@ -44,10 +70,7 @@ export function processTemplate(template: VanillaTemplate): UserTemplate {
 	}
 }
 
-export function transformTemplate(
-	template: VanillaTemplate,
-	version: string
-): BlockObject {
+function transformTemplate(template: VanillaTemplate, version: string): Block {
 	const transformedTemplate = objectMap(
 		template as Required<VanillaTemplate>,
 		(key, value) => {
