+++ packages/atropa/src/core/index.ts
@@ -1,8 +1,6 @@
-export * from './block'
-export * from './item'
-export * from './loot-table'
-export * from './recipe'
-
-export * from './collection'
-export * from './component'
-export * from './hooks'
+export { defineCollection } from './generators/collection'
+export { defineComponent } from './generators/component'
+export { defineBlock } from './generators/block'
+export { defineItem } from './generators/item'
+export { defineLootTable } from './generators/loot-table'
+export { defineRecipe } from './generators/recipe'
