+++ packages/atropa/src/core/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './object'
